Minor refactoring for constants package
