+++ config/random-generator/deployment/random-generator-2.yml
@@ -3,12 +3,13 @@ kind: Deployment
 metadata:
   name: random-generator-2
   annotations:
-    scaler/state-peak-replicas: "5"
+    # scaler/type: autoscale
+    scaler/state-peak-replicas: "50"
     scaler/state-bau-replicas: "2"
     scaler/state-default-replicas: "1"
   labels:
     app: random-generator-2
-    scaler/opt-in: "false"
+    scaler/opt-in: "true"
 spec:
   selector:
     matchLabels:
