+++ internal/validations/label_validation.go
@@ -2,13 +2,8 @@ package validations
 
 import (
 	"errors"
+	c "github.com/containersol/prescale-operator/internal"
 	"strconv"
-)
-
-const LabelNotFound = "Opt-in label was not found"
-
-var (
-	OptInLabel = map[string]string{"scaler/opt-in": "true"}
 )
 
 // OptinLabelExists checks if the opt-in label exists in the target object and returns its value
@@ -16,11 +11,11 @@ func OptinLabelExists(labels map[string]string) (bool, error) {
 
 	var optinlabel bool
 
-	for k := range OptInLabel {
+	for k := range c.OptInLabel {
 		if v, found := labels[k]; found {
 			optinlabel, err := strconv.ParseBool(v)
 			return optinlabel, err
 		}
 	}
-	return optinlabel, errors.New(LabelNotFound)
+	return optinlabel, errors.New(c.LabelNotFound)
 }
